rename function building files
