+++ econpizza/solvers/steady_state.py
@@ -6,7 +6,7 @@ import jax.numpy as jnp
 from copy import deepcopy
 from grgrjax import newton_jax, val_and_jacfwd, amax
 from ..parser import compile_stst_inputs, d2jnp
-from ..parser.build_functions import get_func_stst
+from ..parser.build_generic_functions import get_func_stst
 
 
 def solver(jval, fval):
